supressed warnings due to implicit conversion

+++ src/sim/hands/walker.cc
@@ -27,7 +27,7 @@ void Walker::attach(FiberSite const& s)
     // here digit::step_size must be equal to fiber:step_size
     if ( lattice() && lattice()->unit() != prop->step_size  )
         throw InvalidParameter("digit:step_size must be equal to fiber:lattice_unit");
-    stride = sign_real(prop->unloaded_speed);
+    stride = (int)sign_real(prop->unloaded_speed);
 #endif
 }
